+++ src/test/java/io/lettuce/core/probabilistic/CfInfoValueParserUnitTests.java
@@ -4,7 +4,7 @@
  *
  * SPDX-License-Identifier: MIT
  */
-package io.lettuce.core.cf;
+package io.lettuce.core.probabilistic;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -15,8 +15,6 @@ import java.nio.charset.StandardCharsets;
 import org.junit.jupiter.api.Test;
 
 import io.lettuce.core.output.ComplexData;
-import io.lettuce.core.probabilistic.CfInfoValue;
-import io.lettuce.core.probabilistic.CfInfoValueParser;
 
 /**
  * Unit tests for {@link CfInfoValueParser}.
